requirements.txt
pyserial==3.5
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3